crawler/requirements.txt
flask
python-dotenv
elasticsearch==8.11.0
requests
xxhash
